[SCRIPT] Usage-stats : fix formats.js

+++ usage-stats.js
@@ -10,7 +10,10 @@ const Package = require(Path.resolve(__dirname, 'package.json'));
 
 if(!FileSystem.existsSync('temp'))
 	FileSystem.mkdirSync('temp')
-FileSystem.copyFileSync("node_modules/pokemon-showdown/.config-dist/formats.js",'temp/formats.js')
+if(FileSystem.existsSync("../pokemon-showdown/.config-dist/formats.js"))
+	FileSystem.copyFileSync("../pokemon-showdown/.config-dist/formats.js",'temp/formats.js')	
+else if(FileSystem.existsSync("node_modules/pokemon-showdown/.config-dist/formats.js"))
+	FileSystem.copyFileSync("node_modules/pokemon-showdown/.config-dist/formats.js",'temp/formats.js')	
 /* Globals */
 
 global.toId = function (str) {
